Add workaround for already deleted reference in BaseDTRecordReference

+++ datatype/class.BaseDTRecordReference.php
@@ -366,7 +366,8 @@ abstract class BaseDTRecordReference extends DataType {
 						throw new exception( Debug::getStringRepresentation( $this->value ) . " = \$this->value, not instanceof IRecord!" );
 					}
 
-					$this->value->delete();
+					if (!$this->value->isDeleted())
+						$this->value->delete();
 				}
